+++ marketplace/core/types/channels/whatsapp/type.py
@@ -15,3 +15,12 @@ class WhatsAppType(AppType):
     bg_color = "#d1fcc9cc"  # TODO: Change to real color
     platform = App.PLATFORM_WENI_FLOWS
     config_design = "popup"
+
+    def get_access_token(self, app: App):
+        token = app.config.get("fb_access_token")
+        if token:
+            return token
+
+        raise ValueError(
+            f"not found 'fb_access_token' for app code: {app.code}, app UUID: {str(app.uuid)}"
+        )