+++ marketplace/wpp_templates/tasks.py
@@ -4,8 +4,6 @@ from celery import shared_task
 
 from sentry_sdk import capture_exception
 
-from marketplace.core.types import APPTYPES
-
 from .models import TemplateMessage
 from .requests import TemplateMessageRequest
 
@@ -17,7 +15,7 @@ from marketplace.wpp_templates.models import (
 )
 from marketplace.clients.flows.client import FlowsClient
 
-from .utils import WebhookEventProcessor
+from .utils import WebhookEventProcessor, handle_error_and_update_config
 
 
 logger = logging.getLogger(__name__)
@@ -25,55 +23,80 @@ logger = logging.getLogger(__name__)
 
 @shared_task(track_started=True, name="refresh_whatsapp_templates_from_facebook")
 def refresh_whatsapp_templates_from_facebook():
-    flows_client = FlowsClient()
-
     for app in App.objects.filter(code__in=["wpp", "wpp-cloud"]):
-        if not (app.config.get("wa_waba_id") or app.config.get("waba")):
-            continue
+        try:
+            if not (app.config.get("wa_waba_id") or app.config.get("waba")):
+                continue
 
+            if "ignores_meta_sync" in app.config:
+                logger.info(
+                    f"Skipping sync for app {app.uuid} based on previous error: {app.config['ignores_meta_sync']}"
+                )
+                continue
+
+            service = FacebookTemplateSyncService(app)
+            service.sync_templates()
+
+        except Exception as e:
+            logger.error(f"Error processing app {app.uuid}: {str(e)}")
+
+
+class FacebookTemplateSyncService:
+    def __init__(self, app):
+        self.app = app
+        try:
+            self.client = TemplateMessageRequest(
+                access_token=app.apptype.get_access_token(app)
+            )
+        except ValueError as e:
+            logger.error(f"Access token error for app {app.uuid}: {str(e)}")
+            raise
+
+        self.flows_client = FlowsClient()
+
+    def sync_templates(self):
         waba_id = (
-            app.config.get("waba").get("id")
-            if app.config.get("waba")
-            else app.config.get("wa_waba_id")
+            self.app.config.get("waba").get("id")
+            if self.app.config.get("waba")
+            else self.app.config.get("wa_waba_id")
         )
-        if app.code == "wpp" and app.config.get("fb_access_token"):
-            access_token = app.config.get("fb_access_token")
-        else:
-            access_token = APPTYPES.get("wpp-cloud").get_access_token(app)
 
-        template_message_request = TemplateMessageRequest(access_token=access_token)
-        templates = template_message_request.list_template_messages(waba_id)
+        templates = self.client.list_template_messages(waba_id)
 
         if templates.get("error"):
+            template_error = templates["error"]
             logger.error(
-                f"A error occurred with waba_id: {waba_id}. \nThe error was:  {templates}\n"
+                f"A error occurred with waba_id: {waba_id}. \nThe error was:  {template_error}\n"
             )
-            continue
+            handle_error_and_update_config(self.app, template_error)
+            return
 
         templates = templates.get("data", [])
         try:
-            flows_client.update_facebook_templates(str(app.flow_object_uuid), templates)
+            self.flows_client.update_facebook_templates(
+                str(self.app.flow_object_uuid), templates
+            )
         except Exception as error:
             logger.error(
                 f"An error occurred when sending facebook templates to flows: "
-                f"App-{str(app.uuid)}, flows_object_uuid: {str(app.flow_object_uuid)} "
+                f"App-{str(self.app.uuid)}, flows_object_uuid: {str(self.app.flow_object_uuid)} "
                 f"Error: {error}"
             )
 
         if waba_id:
-            delete_unexistent_translations(app, templates)
+            delete_unexistent_translations(self.app, templates)
 
         for template in templates:
             try:
                 translation = TemplateTranslation.objects.filter(
-                    message_template_id=template.get("id"), template__app=app
+                    message_template_id=template.get("id"), template__app=self.app
                 )
                 if translation:
                     translation = translation.last()
                     found_template = translation.template
                 else:
                     found_template, _created = TemplateMessage.objects.get_or_create(
-                        app=app,
+                        app=self.app,
                         name=template.get("name"),
                     )
 
@@ -135,6 +158,8 @@ def refresh_whatsapp_templates_from_facebook():
                 capture_exception(error)
                 continue
 
+        print(f"Completed template update for app {str(self.app.uuid)}")
+
 
 def delete_unexistent_translations(app, templates):
     templates_message = app.template.all()
@@ -150,13 +175,17 @@ def delete_unexistent_translations(app, templates):
 
             for translation in template_translation:
                 if translation.message_template_id not in templates_ids:
+                    translation_language = (
+                        translation.language if translation.language else "No language"
+                    )
                     print(
-                        f"Removing translation {translation.message_template_id}: {translation}"
+                        f"Removing translation {translation.message_template_id}: {translation_language}"
                     )
                     translation.delete()
 
             if template.translations.all().count() == 0:
-                print(f"Removing template after removing translations: {template}")
+                temaplte_name = template.name if template.name else "No name"
+                print(f"Removing template:{temaplte_name} after removing translations")
                 template.delete()
 
         except Exception as e:
@@ -179,7 +208,7 @@ def update_templates_by_webhook(**kwargs):  # pragma: no cover
             value = change.get("value")
             if field in allowed_event_types:
                 WebhookEventProcessor.process_event(
-                    whatsapp_business_account_id, value, field
+                    whatsapp_business_account_id, value, field, webhook_data
                 )
             else:
                 logger.info(f"Event: {field}, not mapped to usage")