+++ CHANGELOG.md
@@ -1,3 +1,18 @@
+v3.4.6
+----------
+* Updates template per app and ignores apps with errors in waba
+* Fix test_update_template_translation_error test case
+
+v3.4.5
+----------
+* Check that the price and availability fields are filled in correctly
+
+v3.4.4
+----------
+* Facebook webhook
+* Changes in the treatment process and add product validations
+* Remove webhook log info and Add info to log update task
+
 v3.4.3
 ----------
 * Remove saving of vtex products in DB